+++ tests/Factory/ServerRequest/ServerRequestFactoryTest.php
@@ -6,7 +6,6 @@ namespace Tests\Innmind\Http\Factory\ServerRequest;
 use Innmind\Http\{
     Factory\ServerRequestFactory,
     Factory\HeadersFactory,
-    Factory\EnvironmentFactory,
     Factory\CookiesFactory,
     Factory\QueryFactory,
     Factory\FormFactory,
@@ -34,7 +33,6 @@ class ServerRequestFactoryTest extends TestCase
         $f = ServerRequestFactory::of(
             HeadersFactory::of(static fn() => Headers::of()),
             static fn() => Content::none(),
-            EnvironmentFactory::of(static fn() => Environment::of()),
             CookiesFactory::of(static fn() => Cookies::of()),
             QueryFactory::of(static fn() => Query::of([])),
             FormFactory::of(static fn() => Form::of([])),
@@ -58,7 +56,6 @@ class ServerRequestFactoryTest extends TestCase
         $factory = ServerRequestFactory::of(
             HeadersFactory::of(static fn() => Headers::of()),
             static fn() => Content::none(),
-            EnvironmentFactory::of(static fn() => Environment::of()),
             CookiesFactory::of(static fn() => Cookies::of()),
             QueryFactory::of(static fn() => Query::of([])),
             FormFactory::of(static fn() => Form::of([])),
@@ -83,7 +80,6 @@ class ServerRequestFactoryTest extends TestCase
         $factory = ServerRequestFactory::of(
             HeadersFactory::of(static fn() => Headers::of()),
             static fn() => Content::none(),
-            EnvironmentFactory::of(static fn() => Environment::of()),
             CookiesFactory::of(static fn() => Cookies::of()),
             QueryFactory::of(static fn() => Query::of([])),
             FormFactory::of(static fn() => Form::of([])),